Changed long int to long long int

+++ programming/c/control-structures/fibonacci.c
@@ -2,7 +2,7 @@
 
 int main(int argc, char *argv[])
 {
-  long unsigned int f0, f1, f2;
+  long long unsigned int f0, f1, f2;
   int n = 2;
 
     // initialize variables and write the while loop
@@ -11,10 +11,10 @@ int main(int argc, char *argv[])
     f1 = 1;
     f2 = f0 + f1;
 
-    printf("\t0\t%lu\n\t1\t%lu\n",f0,f1);
+    printf("\t0\t%llu\n\t1\t%llu\n",f0,f1);
 
     while(n < 100) {
-      printf("\t%d\t%lu\n",n,f2);
+      printf("\t%d\t%llu\n",n,f2);
       f0 = f1;
       f1 = f2;
       f2 = f0 + f1;
